change app.js server "api/task" to router "cards/get-activity-cards"
