+++ fronted/paintball-app/src/components/calendar.jsx
@@ -21,7 +21,7 @@ const Calendar = () => {
    
     try {
       const response = await fetch(
-        `${process.env.REACT_APP_RENDER_API_URL}/api/tasks`,
+        `${process.env.REACT_APP_RENDER_API_URL}/cards/get-activity-cards`,
         user
           ? {
               method: "GET",
