+++ fronted/paintball-app/src/services/basicIrlAxios.js
@@ -1,6 +1,6 @@
 import axios from "axios";
 import config from "../config.json";
-axios.defaults.baseURL = process.env.REACT_APP_RENDER_API_URL; //   || config.apiUrl;
+axios.defaults.baseURL = process.env.REACT_APP_RENDER_API_URL; 
 export function setCommonHeader(headerName, value) {
   axios.defaults.headers.common[headerName] = value;
 }
